# [DeepSeek] Improve performance of cache kernel