<!-- markdownlint-disable -->
PLEASE FILL IN THE PR DESCRIPTION HERE ENSURING ALL CHECKLIST ITEMS (AT THE BOTTOM) HAVE BEEN CONSIDERED.

## Purpose
Improve the performance of the `concat_and_cache_ds_mla_kernel` by having each thread handle either 8 NoPE elements or 2 RoPE elements.

## Correctness
`python tests/kernels/attention/test_cache.py::test_concat_and_cache_ds_mla` (Passes)


## Performance

### Before

```
concat_and_cache_ds_mla Kernel Benchmark
==================================================
PyTorch version: 2.8.0+cu128
CUDA version: 12.8
GPU: NVIDIA H100 80GB HBM3

Benchmarking concat_and_cache_ds_mla_kernel on cuda:0
Data type: torch.float16
Iterations per test: 1000
================================================================================
Warming up GPU...
GPU warmed up.

Test 1/6: tokens=42, kv_lora_rank=512, qk_rope_head_dim=64, block_size=16, num_blocks=8
  Mean time: 0.012 ± 0.001 ms
  Min/Max time: 0.011 / 0.028 ms
  Throughput: 3411907 tokens/sec, 3.66 GB/s

Test 2/6: tokens=128, kv_lora_rank=512, qk_rope_head_dim=64, block_size=16, num_blocks=32
  Mean time: 0.012 ± 0.001 ms
  Min/Max time: 0.011 / 0.027 ms
  Throughput: 10492036 tokens/sec, 11.26 GB/s

Test 3/6: tokens=256, kv_lora_rank=512, qk_rope_head_dim=64, block_size=16, num_blocks=64
  Mean time: 0.012 ± 0.001 ms
  Min/Max time: 0.012 / 0.025 ms
  Throughput: 20708973 tokens/sec, 22.22 GB/s

Test 4/6: tokens=512, kv_lora_rank=512, qk_rope_head_dim=64, block_size=16, num_blocks=128
  Mean time: 0.013 ± 0.001 ms
  Min/Max time: 0.013 / 0.024 ms
  Throughput: 38085103 tokens/sec, 40.86 GB/s

Test 5/6: tokens=1024, kv_lora_rank=512, qk_rope_head_dim=64, block_size=16, num_blocks=256
  Mean time: 0.015 ± 0.001 ms
  Min/Max time: 0.015 / 0.028 ms
  Throughput: 66554190 tokens/sec, 71.40 GB/s

Test 6/6: tokens=2048, kv_lora_rank=512, qk_rope_head_dim=64, block_size=16, num_blocks=512
  Mean time: 0.020 ± 0.001 ms
  Min/Max time: 0.019 / 0.028 ms
  Throughput: 104694434 tokens/sec, 112.32 GB/s

BENCHMARK SUMMARY
================================================================================
Config                    Mean (ms)    Std (ms)    Tokens/sec   GB/s    
--------------------------------------------------------------------------------
42t,512r,64p              0.012        0.001       3411907      3.66    
128t,512r,64p             0.012        0.001       10492036     11.26   
256t,512r,64p             0.012        0.001       20708973     22.22   
512t,512r,64p             0.013        0.001       38085103     40.86   
1024t,512r,64p            0.015        0.001       66554190     71.40   
2048t,512r,64p            0.020        0.001       104694434    112.32  

Best throughput: 104694434 tokens/sec (config: (2048, 512, 64))
Best latency: 0.012 ms (config: (128, 512, 64))
Worst latency: 0.020 ms (config: (2048, 512, 64))
```

### After

```
concat_and_cache_ds_mla Kernel Benchmark
==================================================
PyTorch version: 2.8.0+cu128
CUDA version: 12.8
GPU: NVIDIA H100 80GB HBM3

Benchmarking concat_and_cache_ds_mla_kernel on cuda:0
Data type: torch.float16
Iterations per test: 1000
================================================================================
Warming up GPU...
GPU warmed up.

Test 1/6: tokens=42, kv_lora_rank=512, qk_rope_head_dim=64, block_size=16, num_blocks=8
  Mean time: 0.012 ± 0.001 ms
  Min/Max time: 0.011 / 0.020 ms
  Throughput: 3549984 tokens/sec, 3.81 GB/s

Test 2/6: tokens=128, kv_lora_rank=512, qk_rope_head_dim=64, block_size=16, num_blocks=32
  Mean time: 0.012 ± 0.001 ms
  Min/Max time: 0.011 / 0.023 ms
  Throughput: 10722024 tokens/sec, 11.50 GB/s

Test 3/6: tokens=256, kv_lora_rank=512, qk_rope_head_dim=64, block_size=16, num_blocks=64
  Mean time: 0.012 ± 0.001 ms
  Min/Max time: 0.012 / 0.021 ms
  Throughput: 21613724 tokens/sec, 23.19 GB/s

Test 4/6: tokens=512, kv_lora_rank=512, qk_rope_head_dim=64, block_size=16, num_blocks=128
  Mean time: 0.012 ± 0.001 ms
  Min/Max time: 0.012 / 0.030 ms
  Throughput: 42152930 tokens/sec, 45.23 GB/s

Test 5/6: tokens=1024, kv_lora_rank=512, qk_rope_head_dim=64, block_size=16, num_blocks=256
  Mean time: 0.012 ± 0.001 ms
  Min/Max time: 0.012 / 0.021 ms
  Throughput: 82408981 tokens/sec, 88.42 GB/s

Test 6/6: tokens=2048, kv_lora_rank=512, qk_rope_head_dim=64, block_size=16, num_blocks=512
  Mean time: 0.013 ± 0.000 ms
  Min/Max time: 0.013 / 0.018 ms
  Throughput: 155499290 tokens/sec, 166.83 GB/s

BENCHMARK SUMMARY
================================================================================
Config                    Mean (ms)    Std (ms)    Tokens/sec   GB/s    
--------------------------------------------------------------------------------
42t,512r,64p              0.012        0.001       3549984      3.81    
128t,512r,64p             0.012        0.001       10722024     11.50   
256t,512r,64p             0.012        0.001       21613724     23.19   
512t,512r,64p             0.012        0.001       42152930     45.23   
1024t,512r,64p            0.012        0.001       82408981     88.42   
2048t,512r,64p            0.013        0.000       155499290    166.83  

Best throughput: 155499290 tokens/sec (config: (2048, 512, 64))
Best latency: 0.012 ms (config: (42, 512, 64))
Worst latency: 0.013 ms (config: (2048, 512, 64))
```

---
<details>
<summary> Essential Elements of an Effective PR Description Checklist </summary>

- [ ] The purpose of the PR, such as "Fix some issue (link existing issues this PR will resolve)".
- [ ] The test plan, such as providing test command.
- [ ] The test results, such as pasting the results comparison before and after, or e2e results
- [ ] (Optional) The necessary documentation update, such as updating `supported_models.md` and `examples` for a new model.
- [ ] (Optional) Release notes update. If your change is user facing, please update the release notes draft in the [Google Doc](https://docs.google.com/document/d/1YyVqrgX4gHTtrstbq8oWUImOyPCKSGnJ7xtTpmXzlRs/edit?tab=t.0).
</details>

**BEFORE SUBMITTING, PLEASE READ <https://docs.vllm.ai/en/latest/contributing>** (anything written below this line will be removed by GitHub Actions)